requirements.txt
pywinauto==0.6.9
pyautogui==0.9.54
opencv-python==4.9.0.80
numpy==1.26.4
pywin32==306
Pillow==10.3.0